refactor(lifecycle): fold breakdown utils into contract
