+++ src/lifecycle-finalize.ts
@@ -1,7 +1,7 @@
 import { estimateTokens } from "./agent-input";
 import type { ChatResponse } from "./api";
 import { t } from "./i18n";
-import type { RunContext } from "./lifecycle-contract";
+import { type RunContext, totalPromptBreakdownTokens } from "./lifecycle-contract";
 import { stripSignalLine } from "./lifecycle-signal";
 import { DISCOVERY_TOOL_SET, READ_TOOL_SET, SEARCH_TOOL_SET, WRITE_TOOL_SET } from "./tool-registry";
 import { scopedCallLog } from "./tool-session";
@@ -15,12 +15,7 @@ export function phaseFinalize(ctx: RunContext): ChatResponse {
         ? t("agent.output.no_response_after_tools")
         : t("agent.output.no_output");
 
-  const promptInputTokens =
-    ctx.promptBreakdownTotals.systemTokens +
-    ctx.promptBreakdownTotals.toolTokens +
-    ctx.promptBreakdownTotals.skillTokens +
-    ctx.promptBreakdownTotals.memoryTokens +
-    ctx.promptBreakdownTotals.messageTokens;
+  const promptInputTokens = totalPromptBreakdownTokens(ctx.promptBreakdownTotals);
   const inputTokens = Math.max(ctx.inputTokensAccum, promptInputTokens);
   const outputTokens = ctx.outputTokensAccum || estimateTokens(output);
