+++ src/test-utils.ts
@@ -12,9 +12,8 @@ import type { Client, PendingState, StreamEvent } from "./client-contract";
 import { createErrorStats } from "./error-handling";
 import { DEFAULT_FEATURE_FLAGS } from "./feature-flags-contract";
 import type { LifecycleDeps } from "./lifecycle";
-import type { RunContext } from "./lifecycle-contract";
+import { createEmptyPromptBreakdownTotals, type RunContext } from "./lifecycle-contract";
 import { defaultLifecyclePolicy } from "./lifecycle-policy";
-import { createEmptyPromptBreakdownTotals } from "./lifecycle-usage";
 import type { Session, SessionState, SessionTokenUsageEntry } from "./session-contract";
 import type { Toolset } from "./tool-registry";
 import { createSessionContext } from "./tool-session";
